The principle of least privilege

+++ app/Http/Controllers/Api/V1/AuthorTicketsController.php
@@ -32,16 +32,13 @@ class AuthorTicketsController extends ApiController
      */
     public function store(StoreTicketRequest $request, $author_id)
     {
-        try {
-            $this->isAble('store', Ticket::class);
-
+        if ($this->isAble('store', Ticket::class)) {
             return new TicketResource(Ticket::create($request->mappedAttributes([
                 'author' => 'user_id'
             ])));
-
-        } catch (AuthorizationException $exception) {
-            return $this->error('You are not authorized to create that resource', 403);
         }
+
+        return $this->error('You are not authorized to create that resource', 403);
     }
 
     /**
@@ -54,59 +51,51 @@ class AuthorTicketsController extends ApiController
                 ->where('user_id', $author_id)
                 ->firstOrFail();
 
-            $this->isAble('delete', $ticket);
-
-            $ticket->delete();
-
-            return $this->ok('Ticket deleted');
-
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $exception) {
+            if ($this->isAble('delete', $ticket)) {
+                $ticket->delete();
+                return $this->ok('Ticket successfully deleted');
+            }
             return $this->error('You are not authorized to delete that resource', 403);
+        } catch (ModelNotFoundException $exception) {
+            return $this->error('Ticket cannot found.', 404);
         }
     }
 
     public function replace(ReplaceTicketRequest $request, $author_id, $ticket_id)
     {
         try {
-
             $ticket = Ticket::where('id', $ticket_id)
                 ->where('user_id', $author_id)
                 ->firstOrFail();
 
-            $this->isAble('replace', $ticket);
+            if ($this->isAble('replace', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
-
-            return new TicketResource($ticket);
+            return $this->error('You are not authorized to update that resource', 403);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $exception) {
-            return $this->error('You are not authorized to replace that resource', 403);
         }
     }
 
     public function update(UpdateTicketRequest $request, $author_id, $ticket_id)
     {
-        // PUT
         try {
             $ticket = Ticket::where('id', $ticket_id)
                 ->where('user_id', $author_id)
                 ->firstOrFail();
 
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
-            return new TicketResource($ticket);
-
-            // TODO: ticket doesn't belong to user
+            return $this->error('You are not authorized to update that resource', 403);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $exception) {
-            return $this->error('You are not authorized to update that resource', 403);
         }
     }
 }
