+++ app/Http/Controllers/Api/V1/UserController.php
@@ -35,14 +35,12 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            $this->isAble('store', User::class);
 
-            return new UserResource(User::create($request->mappedAttributes()));
+            if ($this->isAble('store', User::class)) {
+                return new UserResource(User::create($request->mappedAttributes()));
+            }
 
-        } catch (AuthorizationException $exception) {
             return $this->error('You are not authorized to create that resource',403);
-        }
     }
 
     /**
@@ -63,20 +61,18 @@ class UserController extends ApiController
     public function update(UpdateUserRequest $request, $user_id)
     {
         try {
-
             $user = User::findOrFail($user_id);
 
-            //policy
-            $this->isAble('update', $user);
+            if ($this->isAble('update', $user)) {
+                $user->update($request->mappedAttributes());
 
-            $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-            return new UserResource($user);
+            return $this->error('You are not authorized to update that resource', 403);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $exception) {
-            return $this->error('You are not authorized to update that resource',403);
         }
     }
 
@@ -88,27 +84,31 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            $this->isAble('delete', $user);
+            if ($this->isAble('delete', $user)) {
+                $user->delete();
 
-            $user->delete();
+                return $this->ok('User successfully deleted');
+            }
 
-            return $this->ok('User deleted');
+            return $this->error('You are not authorized to delete that resource', 403);
+
         } catch (ModelNotFoundException $exception) {
-            return $this->error('User cannot be found.', 404);
+            return $this->error('User cannot found.', 404);
         }
     }
 
     public function replace(ReplaceUserRequest $request, $user_id)
     {
         try {
-
             $user = User::findOrFail($user_id);
 
-            $this->isAble('replace', $user);
+            if ($this->isAble('replace', $user)) {
+                $user->update($request->mappedAttributes());
 
-            $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-            return new UserResource($user);
+            return $this->error('You are not authorized to update that resource', 403);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found.', 404);
